requirements.txt
Flask
psycopg2-binary
requests
google-cloud-documentai
python-dotenv
PyPDF2
flask-cors
openai